+++ frontend/src/components/@common/Navbar/index.tsx
@@ -1,6 +1,5 @@
-import { useQueryClient } from '@tanstack/react-query';
 import { useEffect } from 'react';
-import { Link, matchRoutes, useLocation, useNavigate } from 'react-router-dom';
+import { Link, useLocation, useNavigate } from 'react-router-dom';
 import { Button, Roof, Wrapper } from './Navbar.style';
 import useAddToast from 'hooks/@common/useAddToast';
 import useNavbarRoofAnimation from 'hooks/@common/useNavbarRoofAnimation';
@@ -8,25 +7,16 @@ import useCheckSessionId from 'hooks/queries/auth/useCheckSessionId';
 import { URL_PATH } from 'constants/index';
 import NavItem from './NavItem';
 
-const NO_NAVIGATION_BAR_URLS = [
-  URL_PATH.petRegisterForm,
-  URL_PATH.dictDetail,
-  URL_PATH.petEdit,
-  URL_PATH.login,
-  URL_PATH.authorization,
-].map((path) => ({ path }));
-
 const Navbar = () => {
   const navigate = useNavigate();
   const { pathname, state } = useLocation();
-  const queryClient = useQueryClient();
 
   const addToast = useAddToast();
   const { navbarRef, roofPosition, transitionOffset } = useNavbarRoofAnimation(
     state ? state.prevPathname ?? pathname : pathname,
     pathname
   );
-  queryClient.invalidateQueries({ queryKey: ['checkSessionId'] });
+
   const { isSuccess: isLoggedIn } = useCheckSessionId(false);
 
   useEffect(() => {
@@ -49,11 +39,10 @@ const Navbar = () => {
     });
   };
 
-  const hideNavbar = matchRoutes(NO_NAVIGATION_BAR_URLS, pathname) !== null;
   const newHistoryState = { prevPathname: pathname };
 
   return (
-    <Wrapper ref={navbarRef} $hide={hideNavbar}>
+    <Wrapper ref={navbarRef}>
       <Button as={Link} to={URL_PATH.main} state={newHistoryState}>
         <NavItem isActive={roofPosition === 1} iconId="home-line" label="메인" />
       </Button>
